+++ test/unit/messaging/inbound/comms-request/notify-service/try-send-via-notify.test.js
@@ -1,4 +1,5 @@
 import { jest, describe, test, expect, beforeEach } from '@jest/globals'
+
 import mockCommsRequest from '../../../../../mocks/comms-request/v3.js'
 
 const mockSendEmail = jest.fn()
